refactor(graduate): Rules 컴포먼트 이름을 RulesAdminPage로 변경
