+++ apps/graduate/src/pages/admin/rules/styles/RulesAdminPage.css.ts
