+++ apps/graduate/src/routes/_afterLogin/rules.tsx
@@ -1,6 +1,6 @@
 import { createFileRoute } from '@tanstack/react-router';
 
-import { Rules } from '~/pages/admin/rules';
+import { RulesAdminPage } from '~/pages/admin/rules';
 
 export const Route = createFileRoute('/_afterLogin/rules')({
   component: RouteComponent,
@@ -11,5 +11,5 @@ function RouteComponent() {
   if (!auth.isAdmin) {
     return <>유저사이드</>;
   }
-  return <Rules />;
+  return <RulesAdminPage />;
 }
